before revert

+++ USACOClassGold/src/moooo.cpp
@@ -6,14 +6,14 @@
 #include <deque>
 #include <bits/stdc++.h>
 using namespace std;
-const int MAXN = 500001;
+const long MAXN = 500001;
 struct LoudCow{
-    int height;
-    int volume;
+    long height;
+    long volume;
     //bool operator() (LoudCow lc1, LoudCow lc2) {return lc1.pos < lc2.pos;}
 };
 
-int N;
+long N;
 
 bool cmp(LoudCow lc1, LoudCow lc2) 
 { 
@@ -26,23 +26,23 @@ void setIO(string s) {
 }
 
 deque <LoudCow> lcs; 
-int hearingLeft[50001] = {0};
-int hearingRight[50001] = {0};
+long hearingLeft[50001] = {0};
+long hearingRight[50001] = {0};
 LoudCow cows[50001]; 
 //LoudCow Scows[50001];
 
-int solve(){
+long solve(){
     //copy(cows, cows + N, Scows);
     //sort(Scows, Scows + N, &cmp);
     //cout << "Initial Cows in cows array" << endl;
-    /*for(int i = 0; i < N; i ++){
+    /*for(long i = 0; i < N; i ++){
         cout << cows[i].height << endl;
     }*/
     LoudCow maxcow;
     maxcow.height --;
     maxcow.volume --;
-    int maxcowpos = -1;
-    for(int i = 0; i < N; i ++){
+    long maxcowpos = -1;
+    for(long i = 0; i < N; i ++){
         if(cows[i].height > maxcow.height){
             maxcow = cows[i];
             maxcowpos = i;
@@ -53,7 +53,7 @@ int solve(){
     maxcow = LoudCow();
     maxcow.volume --;
     maxcow.height --;
-    for(int i = N - 1; i >= 0; i --){
+    for(long i = N - 1; i >= 0; i --){
         if(cows[i].height > maxcow.height){
             maxcow = cows[i];
             maxcowpos = i;
@@ -61,11 +61,11 @@ int solve(){
         hearingLeft[i] = maxcowpos;
     }
     
-    int volumes[MAXN] = {0};
-    for(int i = 0; i < N; i ++){
-        int curVol = cows[i].volume;
-        int a = hearingLeft[i];
-        int b = hearingRight[i];
+    long volumes[MAXN] = {0};
+    for(long i = 0; i < N; i ++){
+        long curVol = cows[i].volume;
+        long a = hearingLeft[i];
+        long b = hearingRight[i];
         if(a != i){
             volumes[a] += curVol;
         }
@@ -74,12 +74,12 @@ int solve(){
         }
     }
     sort(volumes, volumes + N);
-    /*for(int i = 0; i < N; i ++){
+    /*for(long i = 0; i < N; i ++){
         cout << volumes[i] << " ";
     }
     cout << endl;
     //std::cout << volumes[0] << " and " << volumes[N - 1];
-    for(int i = 0; i < N; i ++){
+    for(long i = 0; i < N; i ++){
         cout << volumes[i] << " ";
     }*/
     return volumes[N - 1];
@@ -94,7 +94,7 @@ int main() {
     //cout << "Test";
     cin >> N;
     //cout<<N<<endl;
-    for(int i = 0; i < N; i ++){
+    for(long i = 0; i < N; i ++){
         cin >> cows[i].height >> cows[i].volume;
         //cout << cows[i].pos <<endl;
     }
